requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altair=5.3.0=pypi_0
asttokens=2.4.1=pypi_0
attrs=23.2.0=pypi_0
blinker=1.8.2=pypi_0
ca-certificates=2024.3.11=haa95532_0
cachetools=5.3.3=pypi_0
certifi=2024.6.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
colorlover=0.3.0=pypi_0
comm=0.2.2=pypi_0
contourpy=1.2.1=pypi_0
cufflinks=0.17.3=pypi_0
cycler=0.12.1=pypi_0
decorator=5.1.1=pypi_0
exceptiongroup=1.2.1=pypi_0
executing=2.0.1=pypi_0
fonttools=4.53.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
idna=3.7=pypi_0
imbalanced-learn=0.12.3=pypi_0
imblearn=0.0=pypi_0
importlib-resources=6.4.0=pypi_0
ipython=8.18.1=pypi_0
ipywidgets=8.1.3=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyterlab-widgets=3.0.11=pypi_0
kiwisolver=1.4.5=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.0=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
numpy=2.0.0=pypi_0
openssl=3.0.14=h827c3e9_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
parso=0.8.4=pypi_0
pillow=10.4.0=pypi_0
pip=24.0=py39haa95532_0
plotly=5.22.0=pypi_0
prompt-toolkit=3.0.47=pypi_0
protobuf=5.27.2=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=16.1.0=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.1.2=pypi_0
python=3.9.19=h1aa4202_1
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.1=pypi_0
referencing=0.35.1=pypi_0
requests=2.32.3=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.1=pypi_0
scikit-learn=1.5.0=pypi_0
scipy=1.13.1=pypi_0
seaborn=0.13.2=pypi_0
setuptools=69.5.1=py39haa95532_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sqlite=3.45.3=h2bbff1b_0
stack-data=0.6.3=pypi_0
streamlit=1.36.0=pypi_0
tenacity=8.4.2=pypi_0
threadpoolctl=3.5.0=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.1=pypi_0
tornado=6.4.1=pypi_0
traitlets=5.14.3=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.2=pypi_0
vc=14.2=h2eaa2aa_4
vs2015_runtime=14.29.30133=h43f2093_4
watchdog=4.0.1=pypi_0
wcwidth=0.2.13=pypi_0
wheel=0.43.0=py39haa95532_0
widgetsnbextension=4.0.11=pypi_0
zipp=3.19.2=pypi_0
